+++ src/classes/MigrationHelper.ts
@@ -4,20 +4,11 @@ export class MigrationHelper {
 
   private systemStore: SystemStore | null = null;
 
-  private readyPromise: Promise<void>;
-
   constructor(private readonly tableName: string) {
-    this.readyPromise = new Promise((resolve) => {
-      void useSystemStore().then((systemStore) => {
-        this.systemStore = systemStore;
-        resolve();
-      });
-    });
+    this.systemStore = useSystemStore();
   }
 
   public async getLastDbVersion(): Promise<number> {
-    await this.readyPromise;
-
     const value = await this.systemStore!.getValue(`lastDbVersion_${this.tableName}`);
     if (value) {
       return Number(value);
@@ -27,7 +18,6 @@ export class MigrationHelper {
   }
 
   public async setLastDbVersion(version: number): Promise<void> {
-    await this.readyPromise;
     await this.systemStore!.setValue(`lastDbVersion_${this.tableName}`, version.toString());
   }