+++ src/components/todos/TodoSidebar.vue
@@ -76,7 +76,7 @@
 
   import { useTodoStore, type Todo } from '../../stores/todoStore';
 
-  const todoStore = await useTodoStore();
+  const todoStore = useTodoStore();
 
   const props = defineProps({
     todoId: {